Interannual variations of atmospheric pCO2 added
If VARIABLE_ANN_ATM_PCO2 is defined in cppdefs.h then the atmospheric
pCO2 concentration is subject to interannual variations. The pCO2
concentration is determined according to the following rules (see the
comments in atm_pco2.F for more details):

- From 1765-1990.5, it is computed from a spline fit to Siple Ice core
  and Mauna Loa data.
- From 1990.5 on, different scenarios can be chosen:
  S350, S450, S550, S650, S750, DS450, DS550, CIS9

The choice of scenario after 1990.5 is done by the 'futr_scen' keyword
in the ASCII input file, e.g.

futr_scen:
   'S750'

Additionally, the point in time when the simulation starts needs to be
specified by the keyword 'start_year', e.g.

start_year:
   2002.5

With this information the atmospheric pCO2 at each time step is computed
by an interpolation in time. The routine get_frac_year (in calendar.F)
is used to transform model time into data time. If USE_REAL_YEAR is
defined, then it is assumed that 1 year = 365 days, otherwise 1 year
is 360 days.

As before, the switch VARIABLE_ATM_PCO2 can be used to switch on a
seasonal cycle for the atmospheric pCO2, independently of
VARIABLE_ANN_ATM_PCO2 and the manner of the interannual variation.

The code in atm_pco2.F was provided by Nicolas Gruber. The files
stab.dat, splco2.dat and cis92a.dat contain the necessary data.