MULT_CLIM_FILE switch and default boundary values for all tracers added
If several climatology files are used, MULT_CLIM_FILE needs to be set
in cppdefs.h (the TWO/THREE/FOUR_CLIM_FILES switches are now obsolete).
In the ASCII input file, the climatology files can then be specified as
follows:

climatology:
   3
   clim_file_1.nc
   clim_file_2.nc
   clim_file_3.nc

So on the first line, the number of clim files is given and then one
filename per line. At most 4 climatology files are possible. Temperature
and salinity must be contained in the file listed first. For each
tracer, all the files are examined in the order in which they appear
in the list. The search terminates as soon as matching data is found.

If none of the climatology files contains data for a tracer, then a
default value will be used for it. Such default values are now
available for each tracer, for both BEC and NPZD.