Fixed two issues related to bgc and physical fluxes in NPZD
1) The routine for writing bgc fluxes was only called correctly
in case the output frequency of the bgc fluxes was equal to that of
the tracer averages. The same was true for the physical fluxes.

2) The averaging phases of bgc and phys fluxes were not adapted after
a restart (as the tracer avg phase is). This meant that it was not
possible to vary the output frequency of bgc or phys fluxes within a run.