DOC: explain how frcv hosts all received variables on all grids
unlike, stated in a previous commit u and v momentum fluxes are
actually defined on the u and v grid BUT hosted in the same
array (frcv) as the fluxes defined on the Rho-grid. Hence some
elements of that array are left uninitialized and unused, which can
get dangerous if the array gets accessed with wrong indices.