ADD:REF: Full rewrite of the OASIS interface
- The interface is now designed to work with arbitrary grids (as
opposed to requiring the exact same phsycal domain and parallel grid
lines), introducing a coupling coefficient and velocity projections.

- The interface was rewriten in a more modular and readable way,
keeping as few cpp statements as possible and moving as much as
possible to F90.

- These changes are incompatible with the previous version as they
rely on an additional ROMSOC auxiliary file and require COSMO to send
both velocity components to ROMS u- and v-points.