ADD:REF: Add an option for sequential coupling and cleanup
The primary goal of this option is to produce OASIS restart files for
a concurent run.

We also cleanup (actually comment out so far) "home-made" code that
was written to generate restart files. The idea is rather to set
options in the namcouple file to export coupling fields and use a
particular time step of these as OASIS restart files.