Allow extrapolation of forcing data
Sometimes messages popped up "current model time is outside of
time bounds", e.g. with the dust and iron forcing from Domitille.
However, it looked like the model time was rather close to one of
the time bounds, so for the Amazon setup, it seemed ok to do an
extrapolation.