FIX: offset in variables read from file
First index in the NetCDF Fortran 90 interface starts at 1. The first
and last points are out of the oasis grid so must be ignored. So
starting at 1 generates an offset between the mask as seen by OASIS
and the actual ROMS mask.

Maybe a cleaner, more flexible, solution than hardcoded "2" is
preferable in case the boundaries must be involved in the
computational domain.