ncjoin: added option for frequency of closing Netcdf files
Option --ncclose was added to ncjoin giving the user the option to
specify the frequency at which all the Netcdf files will be closed
and reopened (in order to prevent ncjoin from allocating more and
more RAM). For example, 'ncjoin --ncclose=5' closes and reopens
the files every 5 time records. The default frequency is 10 time
records.